+++ validator/src/commands/run/mod.rs
@@ -1,5 +1,6 @@
 pub mod args;
 pub mod execute;
+mod xdp;
 
 pub use {args::add_args, execute::execute};